snapshot

+++ core/mesh/mesh_ids.h
@@ -86,7 +86,7 @@ struct MeshIDs_
 
 	static constexpr size_t MAX_MESH_LEVEL = 4;
 
-	static constexpr size_t INDEX_ZERO = 1UL
+	static constexpr long INDEX_ZERO = 1UL
 			<< (INDEX_DIGITS - MAX_MESH_LEVEL - 1);
 
 	static constexpr size_t ID_ZERO = (INDEX_ZERO << (MAX_MESH_LEVEL))
@@ -1203,7 +1203,7 @@ template<size_t N, size_t A> constexpr size_t MeshIDs_<N, A>::_DJ;
 template<size_t N, size_t A> constexpr size_t MeshIDs_<N, A>::_DI;
 template<size_t N, size_t A> constexpr size_t MeshIDs_<N, A>::_DA;
 template<size_t N, size_t A> constexpr size_t MeshIDs_<N, A>::CELL_ID_MASK;
-template<size_t N, size_t A> constexpr size_t MeshIDs_<N, A>::INDEX_ZERO;
+template<size_t N, size_t A> constexpr long MeshIDs_<N, A>::INDEX_ZERO;
 template<size_t N, size_t A> constexpr Real MeshIDs_<N, A>::COORD_ZERO;
 template<size_t N, size_t A> constexpr Real MeshIDs_<N, A>::COORD_TO_INDEX_FACTOR;
 template<size_t N, size_t A> constexpr Real MeshIDs_<N, A>::INDEX_TO_COORD_FACTOR;
